#  sql: save some EvalCtx allocations