EvalCtx is big, so we shouldn't be passing it anyway. Found by digging
into allocations that pprof attributed to the method signature of
sql.GenerateInsertRow. I tried to see if this showed up on any of our
SQL level INSERT benchmarks, but it seemed to be in the noise.

    name                                  old time/op    new time/op    delta
    ConvertToKVs/tpcc/warehouses=1-8         3.83s ± 0%     3.66s ± 1%   -4.53%  (p=0.016 n=4+5)
    ConvertToSSTable/tpcc/warehouses=1-8     5.50s ± 4%     5.13s ± 1%   -6.80%  (p=0.008 n=5+5)

    name                                  old speed      new speed      delta
    ConvertToKVs/tpcc/warehouses=1-8      23.2MB/s ± 0%  24.3MB/s ± 1%   +4.77%  (p=0.016 n=4+5)
    ConvertToSSTable/tpcc/warehouses=1-8  15.1MB/s ± 4%  16.2MB/s ± 1%   +7.27%  (p=0.008 n=5+5)

    name                                  old alloc/op   new alloc/op   delta
    ConvertToKVs/tpcc/warehouses=1-8         953MB ± 0%     704MB ± 0%  -26.17%  (p=0.029 n=4+4)
    ConvertToSSTable/tpcc/warehouses=1-8    1.64GB ± 0%    1.39GB ± 0%  -15.19%  (p=0.016 n=4+5)

    name                                  old allocs/op  new allocs/op  delta
    ConvertToKVs/tpcc/warehouses=1-8         17.4M ± 0%     16.8M ± 0%   -3.45%  (p=0.008 n=5+5)
    ConvertToSSTable/tpcc/warehouses=1-8     17.4M ± 0%     16.8M ± 0%   -3.45%  (p=0.008 n=5+5)

Also a small cleanup in some import code to use EvalCtx's Copy method
instead of passing around a function closure.

Touches #34809

Release note: None